enhance search result page breadcrumb
