+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "BusinessMonitor",
-  "version": "0.1.0+153",
+  "version": "0.1.0+154",
   "private": true,
   "type": "module",
   "scripts": {
